Update README.md
Cryptocurrency exchange platforms have gained immense popularity in recent years, making it one of the fastest-growing sectors in the financial industry. Coinbase is one of the leading cryptocurrency exchange platforms with over 56 million verified users and a market capitalization of $46 billion. With the growing demand for cryptocurrency exchange apps, many entrepreneurs are looking to create their own platforms. However, one of the most significant factors that determine the feasibility of this business idea is the cost of developing such an app.The financial industry has seen a significant surge in popularity of exchange platforms in recent times. Innosoft Group, a top-notch Cryptocurrency Development Company, has played a pivotal role in this trend with their cutting-edge solutions. Coinbase, a leading player in this space, has amassed over 56 million verified users and boasts a market capitalization of $46 billion. As a result, many entrepreneurs are looking to create their own exchange platforms to capitalize on this booming market. However, one of the key factors that determine the viability of such a venture is the development costs involved.